Refactor dom manipulation out of script.js
